pyproject.toml
[project]
name = "autonomize-ai-qa-test-suite"
version = "1.0.0"
description = "Comprehensive test suite for Autonomize AI Agentic Platform - Data Validation & Model Integration"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "Proprietary"}
authors = [
    {name = "Autonomize AI", email = "qa@autonomize.ai"}
]

dependencies = [
    # Core Testing Frameworks
    "pytest>=7.4.3",
    "pytest-html>=4.1.1",
    "pytest-json-report>=1.5.0",
    "pytest-xdist>=3.5.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-timeout>=2.2.0",
    
    # API Testing
    "requests>=2.31.0",
    "httpx>=0.25.2",
    
    # UI Testing (optional - for future implementation if needed)
    # "playwright>=1.40.0",
    # "selenium>=4.15.2",
    
    # Data Validation
    "jsonschema>=4.20.0",
    "pydantic>=2.5.3",
    "cerberus>=1.3.5",
    
    # Healthcare Standards
    "fhir.resources>=7.4.0",
    "hl7>=0.4.7",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "faker>=20.1.0",
    "python-dateutil>=2.8.2",
    
    # Reporting
    "allure-pytest>=2.13.2",
    "junit-xml>=1.9",
    
    # Docker & CI/CD
    "docker>=6.1.3",
    
    # Logging
    "loguru>=0.7.2",
    
    # Security Testing
    "bandit>=1.7.5",
    "safety>=2.3.5",
]

[project.optional-dependencies]
dev = [
    "flake8>=6.0.0",
    "pylint>=2.17.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--html=reports/report.html",
    "--self-contained-html",
    "--json-report",
    "--json-report-file=reports/report.json",
    "--junitxml=reports/junit.xml",
    "--cov=tests",
    "--cov-report=html:reports/coverage",
    "--cov-report=term-missing",
]
markers = [
    "smoke: Smoke tests",
    "regression: Regression tests",
    "agent: Agent integration tests",
    "model: Model integration tests",
    "ui: UI/UX tests",
    "safety: Safety and privacy tests",
    "p0: Critical priority tests",
    "p1: High priority tests",
    "p2: Medium priority tests",
    "p3: Low priority tests",
    "slow: Tests that take longer to execute",
]
timeout = 300

[tool.black]
line-length = 127
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.flake8]
max-line-length = 127
extend-ignore = ["E203", "E266", "E501", "W503"]
max-complexity = 10

